:recycle: refact: 구매내역페이지 변환, myOrderSlice 타입정의
